+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 3.0.0 - 2017-09-22
+
 ### Breaking Changes
 
 -   ANALYTICS-125: Added login option `storeJwt` to store jwt on local file system. **This property will default to `false`**. This was the default behaviour in previous releases.